+++ lib/rblade/helpers/html_string.rb
@@ -1,0 +1,4 @@
+module RBlade
+  class HtmlString
+  end
+end