+++ test/unit/compiler/concerns/compiles_echos_test.rb
@@ -2,11 +2,11 @@ require "test_case"
 
 class CompilesEchoTest < TestCase
   def test_variable
-    assert_compiles_to "{{foo}}", "_out<<h(foo);", "FOO"
-    assert_compiles_to "{{ foo }}", "_out<<h(foo);", "FOO"
+    assert_compiles_to "{{foo}}", "_out<<RBlade.e(foo);", "FOO"
+    assert_compiles_to "{{ foo }}", "_out<<RBlade.e(foo);", "FOO"
     assert_compiles_to "{{
       foo
-      }}", "_out<<h(foo);", "FOO"
+      }}", "_out<<RBlade.e(foo);", "FOO"
 
     assert_compiles_to "{!!foo!!}", "_out<<(foo).to_s;", "FOO"
     assert_compiles_to "{!! foo !!}", "_out<<(foo).to_s;", "FOO"
@@ -16,9 +16,9 @@ class CompilesEchoTest < TestCase
   end
 
   def test_string
-    assert_compiles_to "{{ 'foo' }}", "_out<<h('foo');", "foo"
-    assert_compiles_to "{{ 1 }}", "_out<<h(1);", "1"
-    assert_compiles_to "{{ 'hello dear reader' }}", "_out<<h('hello dear reader');", "hello dear reader"
+    assert_compiles_to "{{ 'foo' }}", "_out<<RBlade.e('foo');", "foo"
+    assert_compiles_to "{{ 1 }}", "_out<<RBlade.e(1);", "1"
+    assert_compiles_to "{{ 'hello dear reader' }}", "_out<<RBlade.e('hello dear reader');", "hello dear reader"
 
     assert_compiles_to "{!! 'foo' !!}", "_out<<('foo').to_s;", "foo"
     assert_compiles_to "{!! 1 !!}", "_out<<(1).to_s;", "1"
@@ -26,10 +26,10 @@ class CompilesEchoTest < TestCase
   end
 
   def test_expression
-    assert_compiles_to "{{ foo + bar << 'BAZ' }}", "_out<<h(foo + bar << 'BAZ');", "FOOBARBAZ"
-    assert_compiles_to %q({{ "foo" + 'bar' }}), %q[_out<<h("foo" + 'bar');], "foobar"
-    assert_compiles_to %q({{ "#{foo}" }}), %q[_out<<h("#{foo}");], "FOO" # standard:disable Lint/InterpolationCheck
-    assert_compiles_to "{{ 'a' * 3 }}", "_out<<h('a' * 3);", "aaa"
+    assert_compiles_to "{{ foo + bar << 'BAZ' }}", "_out<<RBlade.e(foo + bar << 'BAZ');", "FOOBARBAZ"
+    assert_compiles_to %q({{ "foo" + 'bar' }}), %q[_out<<RBlade.e("foo" + 'bar');], "foobar"
+    assert_compiles_to %q({{ "#{foo}" }}), %q[_out<<RBlade.e("#{foo}");], "FOO" # standard:disable Lint/InterpolationCheck
+    assert_compiles_to "{{ 'a' * 3 }}", "_out<<RBlade.e('a' * 3);", "aaa"
 
     assert_compiles_to "{!! foo + bar << 'BAZ' !!}", "_out<<(foo + bar << 'BAZ').to_s;", "FOOBARBAZ"
     assert_compiles_to %q({!! "foo" + 'bar' !!}), %q[_out<<("foo" + 'bar').to_s;], "foobar"
@@ -38,11 +38,11 @@ class CompilesEchoTest < TestCase
   end
 
   def test_multiple
-    assert_compiles_to "{{'foo'}}bar", "_out<<h('foo');_out<<'bar';", "foobar"
-    assert_compiles_to "foo{{'bar'}}", "_out<<'foo';_out<<h('bar');", "foobar"
+    assert_compiles_to "{{'foo'}}bar", "_out<<RBlade.e('foo');_out<<'bar';", "foobar"
+    assert_compiles_to "foo{{'bar'}}", "_out<<'foo';_out<<RBlade.e('bar');", "foobar"
     assert_compiles_to "foo{{
     'bar'
-    }}baz", "_out<<'foo';_out<<h('bar');_out<<'baz';", "foobarbaz"
+    }}baz", "_out<<'foo';_out<<RBlade.e('bar');_out<<'baz';", "foobarbaz"
 
     assert_compiles_to "{!!'foo'!!}bar", "_out<<('foo').to_s;_out<<'bar';", "foobar"
     assert_compiles_to "foo{!!'bar'!!}", "_out<<'foo';_out<<('bar').to_s;", "foobar"
@@ -74,8 +74,8 @@ class CompilesEchoTest < TestCase
   end
 
   def test_dangerous_strings
-    assert_compiles_to %q({{ '"\'\\\\\\'' }}), %q[_out<<h('"\'\\\\\\'');], "&quot;&#39;\\&#39;"
-    assert_compiles_to %q({{ '<&"\'>' }}), %q[_out<<h('<&"\'>');], "&lt;&amp;&quot;&#39;&gt;"
+    assert_compiles_to %q({{ '"\'\\\\\\'' }}), %q[_out<<RBlade.e('"\'\\\\\\'');], "&quot;&#39;\\&#39;"
+    assert_compiles_to %q({{ '<&"\'>' }}), %q[_out<<RBlade.e('<&"\'>');], "&lt;&amp;&quot;&#39;&gt;"
     assert_compiles_to %q(@{{ '"\'\\\\\\'' }}), nil, %q({{ '"\\'\\\\\\'' }})
 
     assert_compiles_to %q({!! '"\'\\\\\\'' !!}), %q[_out<<('"\'\\\\\\'').to_s;], %q("'\\')
@@ -83,20 +83,20 @@ class CompilesEchoTest < TestCase
   end
 
   def test_erb_style
-    assert_compiles_to "<%=foo%>", "_out<<h(foo);", "FOO"
-    assert_compiles_to "<%= foo %>", "_out<<h(foo);", "FOO"
+    assert_compiles_to "<%=foo%>", "_out<<RBlade.e(foo);", "FOO"
+    assert_compiles_to "<%= foo %>", "_out<<RBlade.e(foo);", "FOO"
     assert_compiles_to "<%=
       foo
-      %>", "_out<<h(foo);", "FOO"
+      %>", "_out<<RBlade.e(foo);", "FOO"
   end
 
   def test_limitations
     # The end tag cannot appear within the echo
-    assert_compiles_to "{{ 'foo}}' }}", "_out<<h('foo);_out<<'\\' }}';"
-    assert_compiles_to "<%= 'foo%>' %>", "_out<<h('foo);_out<<'\\' %>';"
+    assert_compiles_to "{{ 'foo}}' }}", "_out<<RBlade.e('foo);_out<<'\\' }}';"
+    assert_compiles_to "<%= 'foo%>' %>", "_out<<RBlade.e('foo);_out<<'\\' %>';"
 
     # A workaround to this is using the alternative syntax
-    assert_compiles_to "<%= 'foo}}' %>", "_out<<h('foo}}');", "foo}}"
-    assert_compiles_to "{{ 'foo%>' }}", "_out<<h('foo%>');", "foo%&gt;"
+    assert_compiles_to "<%= 'foo}}' %>", "_out<<RBlade.e('foo}}');", "foo}}"
+    assert_compiles_to "{{ 'foo%>' }}", "_out<<RBlade.e('foo%>');", "foo%&gt;"
   end
 end